First step, merged core API and HAL, test, works.
